+++ CHANGELOG.md
@@ -13,6 +13,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [1.1.0] - 2025.03.07
+
+### Changed
+
+- Updated dependencies
+- Updated `User` class so admin users inherit all existing permissions for every tenant
+- Updated documentation
+
+### Removed
+
+- Removed `SoftDeletes` trait from the following resource models:
+  - `PermissionsModel`
+  - `TenantInvitationsModel`
+  - `TenantMetaModel`
+  - `TenantRolesModel`
+  - `TenantTeamsModel`
+  - `TenantUserMetaModel`
+  - `UserKeysModel`
+  - `UserMetaModel`
+
 ## [1.0.0] - 2025.01.09
 
 ### Added